+++ packages/@storefront/query/src/search-box/index.ts
@@ -1,10 +1,5 @@
-import { consume, provide, tag, Events, Selectors, Tag } from '@storefront/core';
+import { consume, provide, tag, Events, KEYS, Selectors, Tag } from '@storefront/core';
 import Query from '../query';
-
-const KEY_ENTER = 13;
-const KEY_ESCAPE = 27;
-const KEY_UP = 38;
-const KEY_DOWN = 40;
 
 @consume('query')
 @provide('searchBox')
@@ -17,21 +12,29 @@ class SearchBox {
   };
   state: SearchBox.State = {
     originalQuery: this.select(Selectors.query),
-    onKeyDown: (event) => (event.keyCode === KEY_DOWN || event.keyCode === KEY_UP) && event.preventDefault(),
+    onKeyDown: (event) => [
+      KEYS.DOWN,
+      KEYS.IE_DOWN,
+      KEYS.IE_UP,
+      KEYS.UP
+    ].indexOf(event.key) !== -1 && event.preventDefault(),
     onKeyUp: (event) => {
       event.preventUpdate = true;
-      switch (event.keyCode) {
-        case KEY_ENTER:
+      switch (event.key) {
+        case KEYS.ENTER:
           if (this.services.autocomplete.hasActiveSuggestion()) {
             return this.flux.emit('sayt:select_active');
           } else {
             return this.actions.search(event.target.value);
           }
-        case KEY_ESCAPE:
+        case KEYS.IE_ESCAPE:
+        case KEYS.ESCAPE:
           return this.flux.emit('sayt:hide');
-        case KEY_UP:
+        case KEYS.IE_UP:
+        case KEYS.UP:
           return this.flux.emit('sayt:activate_previous');
-        case KEY_DOWN:
+        case KEYS.IE_DOWN:
+        case KEYS.DOWN:
           return this.flux.emit('sayt:activate_next');
         default:
           const query = event.target.value;