+++ packages/@storefront/structure/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] [patch]
+### Changed
+- SF-1241: Update filtered list component to use table of DOM key strings.
+
 ## [1.47.0] - 2019-01-11
 ### Added
 - SF-1133: Add support for selecting last filtered refinement on Enter.